+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/ChangeLog.md
@@ -8,6 +8,7 @@
 - `Certificate` no longer extends `CertificateProperties`, but instead contains a `CertificateProperties` property named `Properties`.
 - `IssuerBase` has been renamed to `IssuerProperties`.
 - `Issuer` no longer extends `IssuerProperties`, but instead contains a `IssuerProperties` property named `Properties`.
-- 'CertificatePolicy' has been flattened to include all properties from `KeyOptions` and derivative classes.
-- 'KeyOptions' and derivative classes have been removed.
-- 'CertificateKeyType' members have been aligned with 'Azure.Security.KeyVault.Keys.KeyType' members.
+- `CertificatePolicy` has been flattened to include all properties from `KeyOptions` and derivative classes.
+- `KeyOptions` and derivative classes have been removed.
+- `CertificateKeyType` members have been aligned with `Azure.Security.KeyVault.Keys.KeyType` members.
+- `CertificateImport` has been renamed to `CertificateImportOptions`.